Bore freedom, no parent process if notebook
